+++ packages/undead/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.1.2-wip
 
-- Add `ensure_cli_readme_test.dart` to verify CLI `--help` documentation in `README.md`.
+- Adopt centralized `PathFilter` in `UndeadOptions` and CLI for `--exclude` and
+  `--[no-]ignore-generated` file filtering.
+- Add `ensure_cli_readme_test.dart` to verify CLI `--help` documentation in
+  `README.md`.
 
 ## 0.1.1
 
@@ -20,4 +23,3 @@
   (standalone/executables).
 - Custom suppression directives: `// undead:ignore` and
   `// undead:ignore_for_file`.
-